Move around some files in the "database" folder.
