+++ src/lib/UI/layout/components/VacuumComponent.svelte
@@ -1,6 +1,6 @@
 <script lang="ts">
   import { onMount } from "svelte";
-  import { vacuumSoftDeletedEntities } from "$lib/database/manageData";
+  import { vacuumSoftDeletedEntities } from "$lib/database/manageData/manageData";
   import { vacuumDeprecatedLocalStorageKeys } from "$lib/localStorage/vacuum";
 
   // Constants:
