+++ src/lib/database/manageImportExport/parse/recipes.ts
@@ -1,4 +1,4 @@
-import type { ImportedJsonRecipe } from "$lib/database/types/ImportJSON";
+import type { ImportedJsonRecipe } from "$lib/database/manageImportExport/types/ImportJSON";
 import { Recipe, type IRecipe } from "$lib/domain/entities/Recipe";
 import { isNullOrUndefined } from "$lib/helpers/undefinedHelpers";
 import {
